+++ kernel_files/pack_kernel_cl.cl
@@ -21,7 +21,7 @@ __kernel void pack_left_buffer
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t src = 1 + (HALO_DEPTH - column - 1)*2;
+        const int src = 1 + (HALO_DEPTH - column - 1)*2;
         left_buffer[VERT_IDX] = cur_array[THARR2D(src, 0, x_extra)];
     }
 }
@@ -37,7 +37,7 @@ __kernel void unpack_left_buffer
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t dst = 0;
+        const int dst = 0;
         cur_array[THARR2D(dst, 0, x_extra)] = left_buffer[VERT_IDX];
     }
 }
@@ -55,7 +55,7 @@ __kernel void pack_right_buffer
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t src = x_max + x_extra;
+        const int src = x_max + x_extra;
         right_buffer[VERT_IDX] = cur_array[THARR2D(src, 0, x_extra)];
     }
 }
@@ -71,7 +71,7 @@ __kernel void unpack_right_buffer
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t dst = x_max + x_extra + (HALO_DEPTH - column - 1)*2 + 1;
+        const int dst = x_max + x_extra + (HALO_DEPTH - column - 1)*2 + 1;
         cur_array[THARR2D(dst, 0, x_extra)] = right_buffer[VERT_IDX];
     }
 }
@@ -89,7 +89,7 @@ __kernel void pack_bottom_buffer
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t src = 1 + (HALO_DEPTH - row - 1)*2;
+        const int src = 1 + (HALO_DEPTH - row - 1)*2;
         bottom_buffer[HORZ_IDX] = cur_array[THARR2D(0, src, x_extra)];
     }
 }
@@ -105,7 +105,7 @@ __kernel void unpack_bottom_buffer
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t dst = 0;
+        const int dst = 0;
         cur_array[THARR2D(dst, 0, x_extra)] = bottom_buffer[HORZ_IDX];
     }
 }
@@ -123,7 +123,7 @@ __kernel void pack_top_buffer
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t src = y_max + y_extra;
+        const int src = y_max + y_extra;
         top_buffer[HORZ_IDX] = cur_array[THARR2D(0, src, x_extra)];
     }
 }
@@ -139,7 +139,7 @@ __kernel void unpack_top_buffer
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t dst = y_max + y_extra + (HALO_DEPTH - row - 1)*2 + 1;
+        const int dst = y_max + y_extra + (HALO_DEPTH - row - 1)*2 + 1;
         cur_array[THARR2D(0, dst, x_extra)] = top_buffer[HORZ_IDX];
     }
 }
