+++ kernel_files/update_halo_cl.cl
@@ -17,8 +17,8 @@ int depth)
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t src = 1 + (HALO_DEPTH - column - 1)*2;
-        const size_t dst = 0;
+        const int src = 1 + (HALO_DEPTH - column - 1)*2;
+        const int dst = 0;
         cur_array[THARR2D(dst, 0, x_extra)] = y_invert * cur_array[THARR2D(src, 0, x_extra)];
     }
 }
@@ -40,8 +40,8 @@ int depth)
 
     if (row >= HALO_DEPTH - depth && row <= (y_max + HALO_DEPTH - 1) + y_extra + depth)
     {
-        const size_t src = x_max + x_extra;
-        const size_t dst = x_max + x_extra + (HALO_DEPTH - column - 1)*2 + 1;
+        const int src = x_max + x_extra;
+        const int dst = x_max + x_extra + (HALO_DEPTH - column - 1)*2 + 1;
         cur_array[THARR2D(dst, 0, x_extra)] = cur_array[THARR2D(src, 0, x_extra)];
     }
 }
@@ -63,8 +63,8 @@ int depth)
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t src = 1 + (HALO_DEPTH - row - 1)*2;
-        const size_t dst = 0;
+        const int src = 1 + (HALO_DEPTH - row - 1)*2;
+        const int dst = 0;
         cur_array[THARR2D(0, dst, x_extra)] = y_invert * cur_array[THARR2D(0, src, x_extra)];
     }
 }
@@ -86,8 +86,8 @@ int depth)
 
     if (column >= HALO_DEPTH - depth && column <= (x_max + HALO_DEPTH - 1) + x_extra + depth)
     {
-        const size_t src = y_max + y_extra;
-        const size_t dst = y_max + y_extra + (HALO_DEPTH - row - 1)*2 + 1;
+        const int src = y_max + y_extra;
+        const int dst = y_max + y_extra + (HALO_DEPTH - row - 1)*2 + 1;
         cur_array[THARR2D(0, dst, x_extra)] = cur_array[THARR2D(0, src, x_extra)];
     }
 }
